requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
cffi==1.14.0
chardet==3.0.4
discord.py==1.3.3
emoji-country-flag==1.2.1
idna==2.9
motor==2.1.0
multidict==4.7.5
pycparser==2.20
pydantic==1.4
pymongo==3.10.1
PyNaCl==1.3.0
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
websockets==8.1
yarl==1.4.2
